refactor(scalers): rename go module to asya-scalers and regenerate protobuf
Match the module path to the directory/image now that the scaler is a
top-level component (github.com/deliveryhero/asya/asya-scalers, consistent
with asya-sidecar and asya-state-proxy-go). Regenerated the externalscaler
protobuf with protoc v5.29.5 / protoc-gen-go v1.36.11 / protoc-gen-go-grpc
v1.6.1 so the embedded go_package matches; the binary stays scaler-pubsub.
