+++ src/asya-scalers/scaler_test.go
@@ -6,7 +6,7 @@ import (
 	"testing"
 
 	pubsubpb "cloud.google.com/go/pubsub/v2/apiv1/pubsubpb"
-	pb "github.com/deliveryhero/asya/scaler-pubsub/externalscaler"
+	pb "github.com/deliveryhero/asya/asya-scalers/externalscaler"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 )
